CRUD Request

+++ src/main/java/com/Teknisi/controller/TeknisiController.java
@@ -6,8 +6,6 @@ import java.util.Map;
 
 import javax.validation.Valid;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +20,6 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.Teknisi.exception.DataNotfoundException;
 import com.Teknisi.model.Teknisi;
 import com.Teknisi.services.TeknisiService;
 
@@ -30,10 +27,9 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 
-@ApiOperation(value = "/", tags = "Teknisi Profile Controller")
+@ApiOperation(value = "/teknisi", tags = "Teknisi Profile Controller")
 @RestController
 public class TeknisiController {
-	private Logger logger = LoggerFactory.getLogger("TeknisiApplication");
 	
 	@Autowired TeknisiService teknisiService;
 	
@@ -44,7 +40,7 @@ public class TeknisiController {
 			@ApiResponse(code = 403, message = "Forbidden!"),
 			@ApiResponse(code = 404, message = "Not Found")
 	})
-	@RequestMapping(value = "/teknisi", method = RequestMethod.GET)
+	@RequestMapping(value = "/teknisi/showAllTeknisi", method = RequestMethod.GET)
 	public ResponseEntity<Object> retrieveAllTeknisi() {
 		List<Teknisi> listTeknisi = teknisiService.showAllTeknisi();
 		return new ResponseEntity<>(listTeknisi, HttpStatus.OK);
@@ -59,10 +55,14 @@ public class TeknisiController {
 	})
 	@RequestMapping(value = "/teknisi/{id}", method = RequestMethod.GET)
 	public ResponseEntity<Object> retrieveTeknisiById(@Valid @PathVariable("id") Long id) {
-		logger.debug("Get with id : " + id);
-		if(id.equals(null)) throw new DataNotfoundException();
-		Teknisi teknisi = teknisiService.getTeknisiById(id);
-		return new ResponseEntity<>(teknisi, HttpStatus.OK);
+		if(teknisiService.isTeknisiIdExists(id) == true) {
+			Teknisi teknisi = teknisiService.getTeknisiById(id);
+			return new ResponseEntity<>(teknisi, HttpStatus.OK);
+		}else if (teknisiService.isTeknisiIdExists(id) == false ) {
+			return new ResponseEntity<>("Teknisi ID did not exist", HttpStatus.BAD_REQUEST);
+		}else {
+			return new ResponseEntity<>("Teknisi ID cannot be empty", HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	@ApiOperation(value = "Create Teknisi")
@@ -74,13 +74,15 @@ public class TeknisiController {
 	})
 	@RequestMapping(value = "/teknisi/create", method = RequestMethod.POST)
 	public ResponseEntity<Object> createTeknisi(@Valid @RequestBody Teknisi teknisi, final BindingResult bindingResult) {
-		if (teknisi != null && teknisi.getId() > 0) {
+		long id = teknisi.getId();
+		if(teknisiService.isTeknisiIdExists(id) == true) {
 			return new ResponseEntity<>("Teknisi ID already exist", HttpStatus.BAD_REQUEST);
+		}else if (teknisi.getId() == null ) {
+			return new ResponseEntity<>("Teknisi ID cannot be empty", HttpStatus.BAD_REQUEST);
 		}else {
 			teknisiService.insertTeknisi(teknisi);
 			return new ResponseEntity<>("Teknisi Created Successsfully", HttpStatus.OK);
 		}
-		
 	}
 	
 	@ApiOperation(value = "Update Teknisi")
@@ -92,13 +94,15 @@ public class TeknisiController {
 	})
 	@RequestMapping(value = "/teknisi/update", method = RequestMethod.PUT)
 	public ResponseEntity<Object> updateTeknisi(@Valid @RequestBody Teknisi teknisi, final BindingResult bindingResult) {
-		if (teknisi != null && teknisi.getId() > 0) {
+		long id = teknisi.getId();
+		if(teknisiService.isTeknisiIdExists(id) == true) {
 			teknisiService.updateTeknisi(teknisi);
 			return new ResponseEntity<>("Teknisi Updated Successsfully", HttpStatus.OK);
+		}else if (teknisi.getId() == null ) {
+			return new ResponseEntity<>("Teknisi ID cannot be empty", HttpStatus.BAD_REQUEST);
 		}else {
 			return new ResponseEntity<>("Teknisi ID did not exist", HttpStatus.BAD_REQUEST);
 		}
-		
 	}
 	
 	@ApiOperation(value = "Delete Teknisi")
@@ -110,8 +114,14 @@ public class TeknisiController {
 	})
 	@RequestMapping(value = "/teknisi/delete/{id}", method = RequestMethod.DELETE)
 	public ResponseEntity<Object> deleteTeknisi(@Valid @PathVariable("id") Long id, final BindingResult bindingResult) {
-		teknisiService.deleteTeknisiById(id);
-		return new ResponseEntity<>("Teknisi has been deleted", HttpStatus.OK);
+		if(teknisiService.isTeknisiIdExists(id) == true) {
+			teknisiService.deleteTeknisiById(id);
+			return new ResponseEntity<>("Teknisi has been deleted", HttpStatus.OK);
+		}else if (teknisiService.getTeknisiById(id) == null ) {
+			return new ResponseEntity<>("Teknisi ID cannot be empty", HttpStatus.BAD_REQUEST);
+		}else {
+			return new ResponseEntity<>("Teknisi ID did not exist", HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	@ResponseStatus(HttpStatus.BAD_REQUEST)
